refactor(db): abstract chunk operations into own module
